refactor(xpu): move MiniMax H3 RoPE to platform ops
