+++ lightx2v_platform/ops/__init__.py
@@ -24,9 +24,10 @@ elif PLATFORM == "enflame_gcu":
     from .norm.enflame_gcu import *
     from .rope.enflame_gcu import *
 elif PLATFORM == "intel_xpu":
-    # Register platform RMSNorm before the attention modules import the
-    # framework registry and snapshot all PLATFORM_* registries.
-    from .norm.intel_xpu import *
+    # Register platform implementations before the attention modules import
+    # the framework registry and snapshot all PLATFORM_* registries.
+    from .norm.intel_xpu import *  # noqa: I001
+    from .rope.intel_xpu import *
     from .attn.intel_xpu import *
     from .mm.intel_xpu import *
 elif PLATFORM == "iluvatar_cuda":
