+++ lightx2v_platform/ops/rope/intel_xpu/__init__.py
@@ -1,0 +1,1 @@
+from .minimax_h3_rope import MiniMaxH3XpuRope
